:recycle: Move #to_iruby to data_frame_displayable
