+++ test/test_data_frame_displayable.rb
@@ -289,4 +289,91 @@ class DataFrameDisplayableTest < Test::Unit::TestCase
       STR
     end
   end
+
+  sub_test_case 'to_iruby in table mode' do
+    setup do
+      ENV['RED_AMBER_OUTPUT_MODE'] = 'Table'
+    end
+
+    test 'empty' do
+      df = DataFrame.new
+      assert_equal ['text/plain', '(empty DataFrame)'], df.to_iruby
+    end
+
+    test 'simple dataframe' do
+      df = DataFrame.new(x: [1, 2, Float::NAN], y: ['', ' ', nil], z: [true, false, nil])
+      html = 'RedAmber::DataFrame <3 x 3 vectors> <table><tr><th>x</th><th>y</th><th>z</th></tr><tr><td>1.0</td><td>""</td><td>true</td></tr><tr><td>2.0</td><td>" "</td><td>false</td></tr><tr><td>NaN</td><td><i>(nil)</i></td><td><i>(nil)</i></td></tr></table>'
+      assert_equal html, df.to_iruby[1]
+    end
+
+    test 'long dataframe' do
+      df = RedAmber::DataFrame.new(x: [*1..10])
+      html = 'RedAmber::DataFrame <10 x 1 vector> <table><tr><th>x</th></tr><tr><td>1</td></tr><tr><td>2</td></tr><tr><td>3</td></tr><tr><td>4</td></tr><tr><td>&#8942;</td></tr><tr><td>8</td></tr><tr><td>9</td></tr><tr><td>10</td></tr></table>'
+      assert_equal html, df.to_iruby[1]
+    end
+
+    test 'wide dataframe' do
+      raw_record = (1..16).each.with_object({}) { |i, h| h[(64 + i).chr] = [i] }
+      df = RedAmber::DataFrame.new(raw_record)
+      html = 'RedAmber::DataFrame <1 x 16 vectors> <table><tr><th>A</th><th>B</th><th>C</th><th>D</th><th>E</th><th>F</th><th>G</th><th>&#8230;</th><th>J</th><th>K</th><th>L</th><th>M</th><th>N</th><th>O</th><th>P</th></tr><tr><td>1</td><td>2</td><td>3</td><td>4</td><td>5</td><td>6</td><td>7</td><td>&#8230;</td><td>10</td><td>11</td><td>12</td><td>13</td><td>14</td><td>15</td><td>16</td></tr></table>'
+      assert_equal html, df.to_iruby[1]
+    end
+  end
+
+  sub_test_case 'to_iruby in tdr mode' do
+    setup do
+      ENV['RED_AMBER_OUTPUT_MODE'] = 'TDR'
+    end
+
+    test 'empty' do
+      df = DataFrame.new
+      assert_equal ['text/plain', '(empty DataFrame)'], df.to_iruby
+    end
+
+    test 'simple dataframe' do
+      df = DataFrame.new(x: [1, 2, Float::NAN], y: ['', ' ', nil], z: [true, false, nil])
+      html = <<~OUTPUT
+        RedAmber::DataFrame : 3 x 3 Vectors
+        Vectors : 1 numeric, 1 string, 1 boolean
+        # key type    level data_preview
+        0 :x  double      3 [1.0, 2.0, NaN], 1 NaN
+        1 :y  string      3 ["", " ", nil], 1 nil
+        2 :z  boolean     3 [true, false, nil], 1 nil
+      OUTPUT
+      assert_equal html, df.to_iruby[1]
+    end
+
+    test 'long dataframe' do
+      df = RedAmber::DataFrame.new(x: [*1..10])
+      html = <<~OUTPUT
+        RedAmber::DataFrame : 10 x 1 Vector
+        Vector : 1 numeric
+        # key type  level data_preview
+        0 :x  uint8    10 [1, 2, 3, 4, 5, ... ]
+      OUTPUT
+      assert_equal html, df.to_iruby[1]
+    end
+
+    test 'wide dataframe' do
+      raw_record = (1..16).each.with_object({}) { |i, h| h[(64 + i).chr] = [i] }
+      df = RedAmber::DataFrame.new(raw_record)
+      html = <<~OUTPUT
+        RedAmber::DataFrame : 1 x 16 Vectors
+        Vectors : 16 numeric
+        #  key type  level data_preview
+        0  :A  uint8     1 [1]
+        1  :B  uint8     1 [2]
+        2  :C  uint8     1 [3]
+        3  :D  uint8     1 [4]
+        4  :E  uint8     1 [5]
+        5  :F  uint8     1 [6]
+        6  :G  uint8     1 [7]
+        7  :H  uint8     1 [8]
+        8  :I  uint8     1 [9]
+        9  :J  uint8     1 [10]
+         ... 6 more Vectors ...
+      OUTPUT
+      assert_equal html, df.to_iruby[1]
+    end
+  end
 end
